feat(prefer-readonly-types): Merge rules readonly-array and readonly-keyword
